feat: add remove follower operation
